pyproject.toml
[tool.poetry]
name = "my-discord-bot"
version = "0.1.0"
description = ""
authors = ["aleeca <49315610+ryoheinan@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = "^1.7.2"
python-dotenv = "^0.17.1"

[tool.poetry.dev-dependencies]
flake8 = "^3.9.2"
autopep8 = "^1.5.7"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
